fix(env-refactor): merge in environment refactor
